auth + refactoring
