+++ src/controllers/image.controller.js
@@ -1,6 +1,6 @@
 require("dotenv").config();
 const pool = require("./pgpool");
-const jwt_decode = require("jwt-decode");
+const { getIdFromToken } = require("./user.controller");
 
 const getImages = async (req, res) => {
   const response = await pool.query("SELECT * FROM images");
@@ -15,7 +15,7 @@ const getImagesCompact = async (req, res) => {
 };
 
 const getImageById = async (req, res) => {
-  const id = req.params.id;
+  const id = req.params.image_id;
   const response = await pool.query("SELECT * FROM images WHERE id = $1", [id]);
   if (response.rows.length == 0) res.status(404).send("Image not found");
   else res.json(response.rows[0]);
@@ -43,27 +43,24 @@ const updateImageDescription = async (req, res) => {
   const id = req.params.id;
   const description = req.body.description;
   const now = new Date();
-  const response = await pool.query("SELECT * FROM images WHERE id = $1", [id]);
-  if (response.rows.length == 0) res.status(404).send("Image not found");
-  else {
-    await pool.query(
-      "UPDATE images SET description = $1, updated_at = $2 WHERE id = $3",
-      [description, now, id]
-    );
-    res.json({
-      message: "Image Description Updated Succefully",
-    });
-  }
+  await pool.query(
+    "UPDATE images SET description = $1, updated_at = $2 WHERE id = $3",
+    [description, now, id]
+  );
+  res.json({
+    message: "Image Description Updated Succefully",
+  });
 };
 
 const addTagToImage = async (req, res) => {
+  console.log("--------3--------");
   const image_id = req.params.image_id;
   const tag_id = req.params.tag_id;
-  const response = await pool.query(
+  const response2 = await pool.query(
     "SELECT * FROM image_tag WHERE image_id = $1 AND tag_id = $2",
     [image_id, tag_id]
   );
-  if (response.rows.length != 0)
+  if (response2.rows.length != 0)
     res.status(409).send("Image/Tag relationship already exists");
   else {
     await pool.query(
@@ -99,17 +96,12 @@ const removeTagFromImage = async (req, res) => {
 const uploadImage = async (req, res) => {
   const file = req.body.file;
   const description = req.body.description;
-  const tokken = jwt_decode(req.header("authorization").replace("Bearer ", ""));
-  const email = tokken[process.env.EMAIL_URL];
-  const response = await pool.query("SELECT id FROM users WHERE email = $1", [
-    email,
-  ]);
-
+  const user_id = await getIdFromToken(req);
   const now = new Date();
 
   await pool.query(
     "INSERT INTO images (file, description, user_id, created_at, updated_at) VALUES ($1, $2, $3, $4, $4)",
-    [file, description, response.rows[0].id, now]
+    [file, description, user_id, now]
   );
   res.json({
     message: "Image Uploaded Succefully",
@@ -117,13 +109,10 @@ const uploadImage = async (req, res) => {
 };
 
 const deleteImage = async (req, res) => {
-  const id = req.params.id;
+  const id = req.params.image_id;
   const response = await pool.query("SELECT * FROM images WHERE id = $1", [id]);
-  if (response.rows.length == 0) res.status(404).send("Image not found");
-  else {
-    await pool.query("DELETE FROM images WHERE id = $1", [id]);
-    res.json("Image Deleted Succesfully");
-  }
+  await pool.query("DELETE FROM images WHERE id = $1", [id]);
+  res.json("Image Deleted Succesfully");
 };
 
 //--MIDDLEWARE--
@@ -134,6 +123,20 @@ const imageExists = async (req, res, next) => {
   ]);
   if (response.rows.length == 0)
     res.status(404).send("This image doesn't exists");
+  else next();
+};
+
+const imageBelongsToUser = async (req, res, next) => {
+  console.log("--------1--------");
+  const image_id = req.params.image_id;
+  const user_id = await getIdFromToken(req);
+  console.log("--------2--------");
+  const response = await pool.query(
+    "SELECT * FROM images WHERE id = $1 AND user_id = $2",
+    [image_id, user_id]
+  );
+  if (response.rows.length == 0)
+    res.status(409).send("You can't modify this image");
   else next();
 };
 
@@ -149,4 +152,5 @@ module.exports = {
   uploadImage,
   deleteImage,
   imageExists,
+  imageBelongsToUser,
 };
