+++ src/index.js
@@ -4,13 +4,17 @@ const swaggerUi = require("swagger-ui-express");
 const swaggerDocument = require("../swagger.json");
 
 const app = express();
-const port = process.env.PORT || 9000;
+const port = process.env.PORT;
 
 // middlewares
 app.use(express.json());
 
 //routes
-app.use(require("./routes/index.js"));
+app.use(require("./routes/user.routes"));
+app.use(require("./routes/tag.routes"));
+app.use(require("./routes/image.routes"));
+app.use(require("./routes/gallery.routes"));
 
 app.use("/api-docs", swaggerUi.serve, swaggerUi.setup(swaggerDocument));
+
 app.listen(port, () => console.log("Server listening on port: ", port));
